src/watcher_renewal/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-pubsub
google-cloud-firestore
google-cloud-secret-manager
functions-framework
Flask
tenacity
absl-py
google-cloud-datastore